CEX-arbitrage-oppt-scanner/requirements.txt
ccxt==4.2.17
requests==2.31.0
python-dotenv==1.0.0
mypy==1.9.0  # For type checking 